fix: share push availability across entities
