+++ custom_components/nest_yale_lock/lock.py
@@ -378,25 +378,10 @@ class NestYaleLock(NestYaleEntity, LockEntity):
     @property
     def available(self) -> bool:
         """Return if entity is available."""
+        available = self._stream_available()
         age = self._coordinator.last_good_update_age() if hasattr(self._coordinator, "last_good_update_age") else None
-        stale_limit = self._coordinator.hass.data.get(
-                "nest_yale_lock_stale_max",  # optional override
-                getattr(self._coordinator, "_stale_max_seconds", None) or self._default_stale_max(),
-            )
-        if not self._device_data:
-            available = False
-        elif age is None:
-            # If we have data but no timestamp yet, stay available (we seeded above on init)
-            available = True
-        else:
-            available = age < stale_limit
         _LOGGER.debug("Availability check for %s: %s (age=%s)", self._attr_unique_id, available, age)
         return available
-
-    @staticmethod
-    def _default_stale_max() -> int:
-        from .const import STALE_STATE_MAX_SECONDS
-        return STALE_STATE_MAX_SECONDS
 
     @property
     def state(self) -> LockState:
