+++ custom_components/nest_yale_lock/manifest.json
@@ -22,5 +22,5 @@
     "protobuf==6.32.0",
     "PyJWT>=2.8.0,<3.0"
   ],
-  "version": "2026.02.21b35"
+  "version": "2026.02.21b36"
 }
