+++ release-notes/2026.02.21b36.md
@@ -1,0 +1,18 @@
+Fix companion entity availability so Last Action, Auto-Lock, Auto-Lock Duration, and Tamper follow the push stream instead of showing unavailable.
+
+## Highlights
+- Move push-stream availability logic into the shared base entity so all entity types use the same stale-data handling.
+- Keep the main lock and battery sensor aligned with the same shared availability path.
+- Preserve the prior shared protobuf field-name compatibility fixes.
+- Add regression coverage for base-entity availability behavior.
+
+## Why
+- Some companion entities were still inheriting `CoordinatorEntity`'s default availability even after their underlying values decoded correctly.
+- That left entities like Last Action, Auto-Lock, Auto-Lock Duration, and Tamper showing unavailable while the main lock and battery remained available from the live observe stream.
+
+## Validation
+- `./.venv/bin/python -m unittest tests/test_entity.py tests/test_protobuf_handler.py tests/test_api_client.py tests/test_protobuf_compat.py`
+- `./.venv/bin/python -m compileall custom_components/nest_yale_lock tests/test_entity.py`
+
+## Breaking changes
+- None intended. This beta only aligns companion entity availability with the existing push-stream availability model.
